+++ bank-feeds/RELEASES.md
@@ -657,3 +657,13 @@ Based on:
 - [typescript v2.1.0] bank-feeds
 ### Releases
 - [NPM v2.1.0] https://www.npmjs.com/package/@codat/bank-feeds/v/2.1.0 - bank-feeds
+
+## 2023-10-04 00:05:30
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.93.2 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.2.0] bank-feeds
+### Releases
+- [NPM v2.2.0] https://www.npmjs.com/package/@codat/bank-feeds/v/2.2.0 - bank-feeds